+++ offchain/crates/solana-cli/src/command/revenue_distribution/mod.rs
@@ -1,8 +1,6 @@
 mod configure_contributor_rewards;
 mod contributor_rewards;
-mod convert_2z;
 mod fetch;
-mod harvest_2z;
 mod relay;
 mod validator_deposit;
 
@@ -10,7 +8,7 @@ mod validator_deposit;
 
 use std::io::Write;
 
-use anyhow::{Context, Result, ensure};
+use anyhow::{Result, ensure};
 use clap::{Args, Subcommand};
 use doublezero_cli_core::CliContext;
 use doublezero_contributor_rewards::calculator::proof::ShapleyOutputStorage;
@@ -18,22 +16,15 @@ use doublezero_solana_client_tools::{
     account::zero_copy::ZeroCopyAccountOwnedData,
     rpc::{DoubleZeroLedgerConnection, SolanaConnection},
 };
-use doublezero_solana_sdk::{
-    revenue_distribution::{
-        state::{Distribution, SolanaValidatorDeposit},
-        try_is_processed_leaf,
-        types::RewardShare,
-    },
-    sol_conversion::oracle::OraclePriceData,
+use doublezero_solana_sdk::revenue_distribution::{
+    state::{Distribution, SolanaValidatorDeposit},
+    try_is_processed_leaf,
+    types::RewardShare,
 };
 use doublezero_solana_validator_debt::validator_debt::{
     ComputedSolanaValidatorDebt, ComputedSolanaValidatorDebts,
 };
 use solana_sdk::{pubkey::Pubkey, rent::Rent};
-
-// TODO: Add testnet?
-const SOL_2Z_ORACLE_ENDPOINT: &str =
-    "https://sol-2z-oracle-api-v1.mainnet-beta.doublezero.xyz/swap-rate";
 
 #[derive(Debug, Args)]
 pub struct RevenueDistributionCommand {
@@ -54,17 +45,7 @@ pub enum RevenueDistributionSubcommand {
     #[command(name = "configure-contributor-rewards")]
     ConfigureContributorRewards(configure_contributor_rewards::ConfigureContributorRewardsCommand),
 
-    /// Using the Revenue Distribution program's SOL liquidity, convert 2Z
-    /// tokens to SOL. If there is not enough SOL liquidity for the
-    /// fixed-quantity conversion, the command will fail.
-    #[command(name = "convert-2z")]
-    Convert2z(convert_2z::Convert2zCommand),
-
-    #[command(name = "harvest-2z")]
-    Harvest2z(harvest_2z::Harvest2zCommand),
-
-    /// Manage a Solana validator deposit account. Funding can be directly with
-    /// SOL or with 2Z limited by specified conversion rate for 2Z -> SOL.
+    /// Manage a Solana validator deposit account.
     ValidatorDeposit(validator_deposit::ValidatorDepositCommand),
 
     /// Relayer instructions for the Revenue Distribution program.
@@ -77,8 +58,6 @@ impl RevenueDistributionSubcommand {
             Self::Fetch(command) => command.execute(ctx, out).await,
             Self::ContributorRewards(command) => command.execute(ctx, out).await,
             Self::ConfigureContributorRewards(command) => command.execute(ctx, out).await,
-            Self::Convert2z(command) => command.execute(ctx, out).await,
-            Self::Harvest2z(command) => command.execute(ctx, out).await,
             Self::ValidatorDeposit(command) => command.execute(ctx, out).await,
             Self::Relay(command) => command.inner.execute(ctx, out).await,
         }
@@ -132,18 +111,6 @@ async fn try_fetch_solana_validator_deposit(
         }
         Err(_) => Ok((solana_validator_deposit_key, None, 0)),
     }
-}
-
-async fn try_request_oracle_conversion_price() -> Result<OraclePriceData> {
-    reqwest::Client::new()
-        .get(SOL_2Z_ORACLE_ENDPOINT)
-        .header("User-Agent", "DoubleZero Solana CLI")
-        .send()
-        .await
-        .with_context(|| format!("Failed to request SOL/2Z price from {SOL_2Z_ORACLE_ENDPOINT}"))?
-        .json()
-        .await
-        .context("Failed to parse oracle response. Please try again")
 }
 
 async fn try_fetch_shapley_record(